Switch to using +load instead of __attribute__(constructor)
Summary:
Because +load are called before static initializers and __attribute__ it could result in crashes when asserting that a context exists if a +load method were to try to log messages. We now default to register and create context in a +load method to guarantee a context always exists before it is needed for logging.

- Use AUTLogContextIdentifier instead of NSInteger
- Minor alignment
- Switch to a class +load method
- Fix typo in macro

Version: major

Test Plan: Unit Tests

Reviewers: blossom, engin, eric, James

Reviewed By: eric, James

Subscribers: jenkins

Differential Revision: https://phabricator.automatic.co/D9808